fix: move command code-samples into sub-folders
