+++ code-samples/command-handling/adding-features/11/commands/fun/beep.js
