+++ code-samples/command-handling/adding-features/11/commands/fun/ping.js
