+++ code-samples/command-handling/adding-features/11/commands/moderation/kick.js
