+++ code-samples/command-handling/adding-features/11/commands/moderation/prune.js
