+++ code-samples/command-handling/adding-features/11/commands/utility/avatar.js
