+++ code-samples/command-handling/adding-features/12/commands/fun/ping.js
