+++ code-samples/command-handling/adding-features/12/commands/moderation/prune.js
