+++ code-samples/command-handling/adding-features/12/commands/utility/avatar.js
