+++ guide/command-handling/adding-features.md
@@ -31,7 +31,7 @@ In this short (but necessary) refactor, you:
 
 Now you can start adding features!
 
-## Command Categories
+## Command categories
 
 So far all your command files are in a single `commands` folder. This is fine if you only have a few of them but if you're planning to make a bot that does lots of stuff you might end up with a large number of files in the `commands` folder. Keeping track of that many files can be a little tough. To make this a little easier, you can sort your commands according to their types and put them in separate folders inside the `commands` folder. You will have to make a few changes to your existing code in `index.js` for this to work out.
 
@@ -327,7 +327,7 @@ module.exports = {
 You're going to need your prefix variable a couple times inside this command, so make sure to require that at the very top of the file (outside of the `module.exports` bit).
 
 ```js
-const { prefix } = require('../config.json');
+const { prefix } = require('../../config.json');
 ```
 
 Inside the `execute()` function, set up some variables and an if/else statement to determine whether it should display a list of all the command names or only information about a specific command.
